--- a/l10n_br_account/models/account_invoice.py
+++ b/l10n_br_account/models/account_invoice.py
@@ -251,10 +251,17 @@ class AccountMove(models.Model):
 
     @api.model_create_multi
     def create(self, vals_list):
-        recompute_inv_after_id = self.search([], order="id DESC", limit=1).id
+        recompute_account_move_after_id = self.search([], order="id DESC", limit=1).id
+        recompute_account_move_line_after_id = (
+            self.env["account.move.line"].search([], order="id DESC", limit=1).id
+        )
+
         invoices = super(
             AccountMove,
-            self.with_context(recompute_account_move_after_id=recompute_inv_after_id),
+            self.with_context(
+                recompute_account_move_after_id=recompute_account_move_after_id,
+                recompute_account_move_line_after_id=recompute_account_move_line_after_id,
+            ),
         ).create(vals_list)
         invoices._write_shadowed_fields()
         return invoices