--- a/l10n_br_account/models/fiscal_document_line.py
+++ b/l10n_br_account/models/fiscal_document_line.py
@@ -12,3 +12,21 @@ class FiscalDocumentLine(models.Model):
         inverse_name="fiscal_document_line_id",
         string="Invoice Lines",
     )
+
+    def modified(self, fnames, create=False, before=False):
+        """
+        Modifying a dummy fiscal document line should not recompute
+        any account.move.line related to the same dummy fiscal document line.
+        """
+        if not self.document_id.document_type_id:
+            return
+        return super().modified(fnames, create, before)
+
+    def _modified_triggers(self, tree, create=False):
+        """
+        Modifying a dummy fiscal document line should not recompute
+        any account.move.line related to the same dummy fiscal document line.
+        """
+        if not self.document_id.document_type_id:
+            return []
+        return super()._modified_triggers(tree, create)