--- a/l10n_br_nfe/models/document_line.py
+++ b/l10n_br_nfe/models/document_line.py
@@ -304,7 +304,7 @@ class NFeLine(spec_models.StackedModel):
         if not self.cofinsst_value:
             xsd_fields.remove("nfe40_COFINSST")
 
-        if not self.ii_value:
+        if not self.ii_value and "nfe40_II" in xsd_fields:
             xsd_fields.remove("nfe40_II")
 
     ##################################################