mesmo PR do que https://github.com/OCA/l10n-brazil/pull/2419 mas com commit() depois de criar cada invoice. Por iser, sem commit, esse teste https://github.com/OCA/l10n-brazil/blob/14.0/l10n_br_account/__init__.py#L26 não funciona (os registros ainda não tem id no banco, são de tipo NewId) e os registros se acumulam, pelo menos nos browse e isso parece deixar o loop de criação mais lento com o crescimento dos registros. Eu não sei bem se teria como fazer um teste mais inteligente com os registros NewId e que ainda seria seguro...

Mas num uso real em produção o que importa é de não sofrer impacto com os milhões de account.move.line presente no banco, ja não seria tão grave se na hora de criar 10000 account.move.line de uma so vez num loop (sem usar o create_multi e sem dar commit) fique 4x mais lento na hora de criar o ultimo registro.

Vamos testar isso agora dando commit dentro do loop agora, o que vai sim deixar tudo mais lento no inicio, mas deve ficar melhor depois por não acumular recomputes.

começa um pouco mais lento, mas é normal, agora tem commit tb:

```
2023-04-05 07:10:44,226 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 0 of 1000. time: 0.401187 seconds 
2023-04-05 07:10:44,610 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 1 of 1000. time: 0.366470 seconds 
2023-04-05 07:10:44,976 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 2 of 1000. time: 0.356945 seconds 
2023-04-05 07:10:45,350 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 3 of 1000. time: 0.363472 seconds 
2023-04-05 07:10:45,721 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 4 of 1000. time: 0.362022 seconds 
2023-04-05 07:10:46,084 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 5 of 1000. time: 0.353853 seconds 
2023-04-05 07:10:46,456 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 6 of 1000. time: 0.361826 seconds 
2023-04-05 07:10:46,820 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 7 of 1000. time: 0.353877 seconds 
2023-04-05 07:10:47,189 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 8 of 1000. time: 0.358451 seconds 
2023-04-05 07:10:47,549 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 9 of 1000. time: 0.351189 seconds 
2023-04-05 07:10:47,914 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 10 of 1000. time: 0.354577 seconds 
2023-04-05 07:10:48,261 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 11 of 1000. time: 0.337271 seconds 
2023-04-05 07:10:48,612 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 12 of 1000. time: 0.337729 seconds 
2023-04-05 07:10:48,984 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 13 of 1000. time: 0.360112 seconds 
2023-04-05 07:10:49,353 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 14 of 1000. time: 0.358579 seconds 
2023-04-05 07:10:49,721 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 15 of 1000. time: 0.358519 seconds 
2023-04-05 07:10:50,086 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 16 of 1000. time: 0.355349 seconds 
2023-04-05 07:10:50,452 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 17 of 1000. time: 0.355538 seconds 
2023-04-05 07:10:50,812 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 18 of 1000. time: 0.350281 seconds 
2023-04-05 07:10:51,161 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 19 of 1000. time: 0.339442 seconds 
2023-04-05 07:10:51,534 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 20 of 1000. time: 0.359532 seconds 
2023-04-05 07:10:51,881 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 21 of 1000. time: 0.337742 seconds 
2023-04-05 07:10:52,252 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 22 of 1000. time: 0.359367 seconds 
2023-04-05 07:10:52,619 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 23 of 1000. time: 0.356382 seconds 
2023-04-05 07:10:52,967 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 24 of 1000. time: 0.338361 seconds 
2023-04-05 07:10:53,319 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 25 of 1000. time: 0.338650 seconds 
2023-04-05 07:10:54,418 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 28 of 1000. time: 0.366559 seconds 
2023-04-05 07:10:54,768 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 29 of 1000. time: 0.339810 seconds 
2023-04-05 07:10:55,141 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 30 of 1000. time: 0.360207 seconds 

[...]

2023-04-05 07:16:30,995 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 970 of 1000. time: 0.344889 seconds 
2023-04-05 07:16:31,354 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 971 of 1000. time: 0.344613 seconds 
2023-04-05 07:16:31,708 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 972 of 1000. time: 0.340641 seconds 
2023-04-05 07:16:32,067 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 973 of 1000. time: 0.346160 seconds 
2023-04-05 07:16:32,429 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 974 of 1000. time: 0.348397 seconds 
2023-04-05 07:16:32,783 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 975 of 1000. time: 0.339447 seconds 
2023-04-05 07:16:33,144 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 976 of 1000. time: 0.348642 seconds 
2023-04-05 07:16:33,498 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 977 of 1000. time: 0.340988 seconds 
2023-04-05 07:16:33,858 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 978 of 1000. time: 0.346060 seconds 
2023-04-05 07:16:34,213 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 979 of 1000. time: 0.342323 seconds 
2023-04-05 07:16:34,562 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 980 of 1000. time: 0.335278 seconds 
2023-04-05 07:16:34,919 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 981 of 1000. time: 0.343861 seconds 
2023-04-05 07:16:35,273 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 982 of 1000. time: 0.341076 seconds 
2023-04-05 07:16:35,633 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 983 of 1000. time: 0.346281 seconds 
2023-04-05 07:16:35,992 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 984 of 1000. time: 0.345970 seconds 
2023-04-05 07:16:36,350 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 985 of 1000. time: 0.344736 seconds 
2023-04-05 07:16:36,703 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 986 of 1000. time: 0.339033 seconds 
2023-04-05 07:16:37,056 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 987 of 1000. time: 0.339978 seconds 
2023-04-05 07:16:37,416 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 988 of 1000. time: 0.346008 seconds 
2023-04-05 07:16:38,493 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 991 of 1000. time: 0.346582 seconds 
2023-04-05 07:16:38,850 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 992 of 1000. time: 0.344575 seconds 
2023-04-05 07:16:39,205 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 993 of 1000. time: 0.340213 seconds 
2023-04-05 07:16:39,566 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 994 of 1000. time: 0.348050 seconds 
2023-04-05 07:16:39,925 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 995 of 1000. time: 0.344597 seconds 
2023-04-05 07:16:40,289 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 996 of 1000. time: 0.350666 seconds 
2023-04-05 07:16:40,648 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 997 of 1000. time: 0.344571 seconds 
2023-04-05 07:16:41,008 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 998 of 1000. time: 0.346842 seconds 
2023-04-05 07:16:41,366 301 INFO odoo odoo.addons.l10n_br_account.hooks: Created the invoice 999 of 1000. time: 0.344717 seconds 
```

mas ai pelo menos depois o tempo é constante como pode ser visto. Problem solved.
pera ai, vou passar meu PIX ;-)

cc @antoniospneto @felipemotter @marcelsavegnago @renatonlima