@rvalyi 2023-04-05
eh bom vc testar, pode usar em produção TB se vc precisa. Mas depois vou fazer um PR sem o gerador de teste e vou ver se eu consigo simplificar talvez o fix que eu tinha feito a semana passada, ou se der para evitar todos recomputes ou browser mesmo num loop com todos invoices na memória como o PR inicial do @antoniospneto .

Mas mesmo se ainda tiver essa limitação seria de boa porque vc nunca cria 10 000 account.move.line num loop sem dar commit no meio (no mínimo mesmo se tiver um job assim vc daria commit todos x registros ate para evitar erros ligados a "transaction isolation cursor") e até se vc fizer ainda leva menos de 1 secondo para criar um invoice de 10 linhas o que ainda tá bem razoável.