@rvalyi 2023-04-09
eu dei um revert no primeiro fix porque o segundo ja foi suficiente para matar completamente o problema, vcs podem verificar nos logs que os tempos ficaram igual ao que eu falei no primeiro comentário (tempo constante).